+++ internal/protocol/generated.go
@@ -47,6 +47,14 @@ type ApprovePRMessage struct {
 	Repo string `json:"repo"`
 }
 
+type AuthorState struct {
+	// Author corresponds to the JSON schema field "author".
+	Author string `json:"author"`
+
+	// Muted corresponds to the JSON schema field "muted".
+	Muted bool `json:"muted"`
+}
+
 type Branch struct {
 	// CommitHash corresponds to the JSON schema field "commit_hash".
 	CommitHash *string `json:"commit_hash,omitempty"`
@@ -685,6 +693,14 @@ type MergePRMessage struct {
 	Repo string `json:"repo"`
 }
 
+type MuteAuthorMessage struct {
+	// Author corresponds to the JSON schema field "author".
+	Author string `json:"author"`
+
+	// Cmd corresponds to the JSON schema field "cmd".
+	Cmd string `json:"cmd"`
+}
+
 type MuteMessage struct {
 	// Cmd corresponds to the JSON schema field "cmd".
 	Cmd string `json:"cmd"`
@@ -712,6 +728,9 @@ type MuteRepoMessage struct {
 type PR struct {
 	// ApprovedByMe corresponds to the JSON schema field "approved_by_me".
 	ApprovedByMe bool `json:"approved_by_me"`
+
+	// Author corresponds to the JSON schema field "author".
+	Author string `json:"author"`
 
 	// CIStatus corresponds to the JSON schema field "ci_status".
 	CIStatus *string `json:"ci_status,omitempty"`
@@ -817,6 +836,11 @@ type PRVisitedMessage struct {
 	ID string `json:"id"`
 }
 
+type QueryAuthorsMessage struct {
+	// Cmd corresponds to the JSON schema field "cmd".
+	Cmd string `json:"cmd"`
+}
+
 type QueryMessage struct {
 	// Cmd corresponds to the JSON schema field "cmd".
 	Cmd string `json:"cmd"`
@@ -879,6 +903,8 @@ type RecentLocationsResultMessage struct {
 	// Success corresponds to the JSON schema field "success".
 	Success bool `json:"success"`
 }
+
+type RecordString map[string]interface{}
 
 type RefreshPRsMessage struct {
 	// Cmd corresponds to the JSON schema field "cmd".
@@ -970,6 +996,9 @@ type ResolveCommentResultMessage struct {
 }
 
 type Response struct {
+	// Authors corresponds to the JSON schema field "authors".
+	Authors []AuthorState `json:"authors,omitempty"`
+
 	// Error corresponds to the JSON schema field "error".
 	Error *string `json:"error,omitempty"`
 
@@ -1331,6 +1360,9 @@ type UpdateCommentResultMessage struct {
 }
 
 type WebSocketEvent struct {
+	// Authors corresponds to the JSON schema field "authors".
+	Authors []AuthorState `json:"authors,omitempty"`
+
 	// Branch corresponds to the JSON schema field "branch".
 	Branch *string `json:"branch,omitempty"`
 
@@ -1377,7 +1409,7 @@ type WebSocketEvent struct {
 	Sessions []Session `json:"sessions,omitempty"`
 
 	// Settings corresponds to the JSON schema field "settings".
-	Settings map[string]interface{} `json:"settings,omitempty"`
+	Settings RecordString `json:"settings,omitempty"`
 
 	// StashRef corresponds to the JSON schema field "stash_ref".
 	StashRef *string `json:"stash_ref,omitempty"`